refactor: 서버/클라이언트 컴포넌트 분리 및 타입 에러 수정
