+++ src/app/interview/[id]/page.tsx
@@ -1,28 +1,10 @@
 // 면접 상세 페이지
-'use client';
-
-import { useState } from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
-import ApplicantCardBasic from '@/components/interview/ApplicantCardBasic';
+import InterviewDetailClient from '@/components/interview/InterviewDetailClient';
 import QuestionSection from '@/components/document/QuestionSection';
-import AppointmentModal from '@/components/interview/AppointmentModal';
 
 export default function ApplicantDetailPage() {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [appointmentDate, setAppointmentDate] = useState('11월 15일 (수)');
-  const [appointmentTime, setAppointmentTime] = useState('14:00');
-
-  const handleAppointmentClick = () => {
-    setIsModalOpen(true);
-  };
-
-  const handleConfirm = (date: string, time: string) => {
-    setAppointmentDate(date);
-    setAppointmentTime(time);
-    setIsModalOpen(false);
-  };
-
   return (
     <div className="min-h-screen bg-gray-50 pb-20">
       {/* 상세 페이지 헤더 */}
@@ -36,22 +18,10 @@ export default function ApplicantDetailPage() {
         </button>
       </header>
 
-      <div className="p-4">
-        {/* 상세 내용 */}
-        <ApplicantCardBasic
-          name="김민준"
-          gender="남"
-          status="합격"
-          university="건국대/컴퓨터공학과/일반부원"
-          phone="010-1234-5678"
-          email="minjun@gmail.com"
-          commentCount={3}
-          appointmentDate={appointmentDate}
-          appointmentTime={appointmentTime}
-          onAppointmentClick={handleAppointmentClick}
-        />
-      </div>
+      {/* 클라이언트 컴포넌트 - ApplicantCardBasic과 AppointmentModal */}
+      <InterviewDetailClient />
 
+      {/* 질문 섹션 - 정적 콘텐츠 */}
       <div className="p-4 mx-4 bg-white rounded-10">
         <QuestionSection
           title="지원동기"
@@ -66,15 +36,6 @@ export default function ApplicantDetailPage() {
 
         <QuestionSection title="질문내용" content="질문답변" />
       </div>
-
-      {/* 모달 */}
-      <AppointmentModal
-        isOpen={isModalOpen}
-        onClose={() => setIsModalOpen(false)}
-        onConfirm={handleConfirm}
-        initialDate={appointmentDate}
-        initialTime={appointmentTime}
-      />
     </div>
   );
 }
